refactor: move TrackEvent to new file
